hw04-auth

+++ repositories/contacts.js
@@ -1,41 +1,63 @@
-const Contact = require("../model/contacts"),
+const Contact = require('../model/contacts'),
 
-const getAll = async () => {
-  const results = await Contact.find(),
-  return results,
-},
+const getAll = async (userId, query) => {
+  const { limit = 20, offset = 0 } = query
+  const optionSearch = { owner: userId }
+  const results = await Contact.paginate(optionsSearch, {
+    limit, offset,
+    populate: {
+      path: 'owner',
+      select: 'name email subscription',
+    }
+  })
+  return results
+}
 
-const getById = async (contactId) => {
-  const result = await Contact.findOne({ _id: contactId }),
-  return result,
-},
+const getById = async (userId, contactId) => {
+  const result = await Contact.findOne({ _id: contactId, owner: userId
+  }).populate({
+    path: 'owner',
+    select: 'name email subscription'
+  })
+  return result
+}
 
 const removeContact = async (contactId) => {
-  const result = await Contact.findOneAndDelete({ _id: contactId }),
-  return result,
+  const result = await Contact.findOneAndDelete({ _id: contactId, owner: userId
+  }).populate({
+    path: 'owner',
+  select: 'name email subscription'
+})
+  return result
+}
+
+const addContact = async (userId, body) => {
+  const result = await Contact.create(owner: userId, ...body),
+  return result
 },
 
-const addContact = async (body) => {
-  const result = await Contact.create(body),
-  return result,
-},
-
-const updateContact = async (contactId, body) => {
+const updateContact = async (userId, contactId, body) => {
   const result = await Contact.findOneAndUpdate(
-    { _id: contactId },
+    { _id: contactId, owner: userId },
     { ...body },
     { new: true }
-  ),
-  return result,
+  ).populate({
+    path: 'owner',
+    select: 'name email phone',
+  })
+  return result
 },
 
 const updateStatusContact = async (contactId, body) => {
   const result = await Contact.findOneAndUpdate(
-    { _id: contactId },
+    { _id: contactId, owner: userId },
     { ...body },
     { new: true }
-  ),
-  return result,
+  ).populate({
+    path: 'owner',
+    select: 'name email phone',
+  })
+  return result
 },
 
 module.exports = {
